+++ R/learner_catboost_classif_catboost.R
@@ -30,6 +30,11 @@
 #'   - Adjusted default: FALSE
 #'   - Reason for change: consistent with other mlr3 learners
 #'
+#' @section Early stopping:
+#' Early stopping can be used to find the optimal number of boosting rounds.
+#' Set `early_stopping_rounds` to an integer value to monitor the performance of the model on the validation set while training.
+#' For information on how to configure the validation set, see the *Validation* section of [`mlr3::Learner`].
+#'
 #' @references
 #' `r format_bib("dorogush2018catboost")`
 #'
@@ -53,8 +58,6 @@ LearnerClassifCatboost = R6Class("LearnerClassifCatboost",
         loss_function_multiclass = p_fct(levels = c("MultiClass", "MultiClassOneVsAll"),
           default = "MultiClass", tags = "train"),
         # custom_loss missing
-        # eval_metric missing
-        iterations = p_int(lower = 1L, upper = Inf, default = 1000, tags = "train"),
         learning_rate = p_dbl(lower = 0.001, upper = 1, default = 0.03, tags = "train"),
         random_seed = p_int(lower = 0, upper = Inf, default = 0, tags = "train"),
         l2_leaf_reg = p_dbl(lower = 0, upper = Inf, default = 3, tags = "train"),
@@ -147,7 +150,25 @@ LearnerClassifCatboost = R6Class("LearnerClassifCatboost",
         # https://catboost.ai/docs/concepts/r-reference_catboost-predict.html
         verbose = p_lgl(default = FALSE, tags = "predict"),
         ntree_start = p_int(lower = 0L, upper = Inf, default = 0L, tags = "predict"),
-        ntree_end = p_int(lower = 0L, upper = Inf, default = 0L, tags = "predict")
+        ntree_end = p_int(lower = 0L, upper = Inf, default = 0L, tags = "predict"),
+        # early stopping
+        early_stopping_rounds = p_int(lower = 1L, upper = Inf, tags = "train"),
+        eval_metric = p_uty(tags = "train"),
+        use_best_model = p_lgl(tags = "train"),
+        iterations = p_int(
+          lower = 1L,
+          upper = Inf,
+          default = 1000L,
+          tags = c("train", "internal_tuning"),
+          aggr = crate(function(x) as.integer(ceiling(mean(unlist(x)))), .parent = topenv()),
+          in_tune_fn = crate(function(domain, param_vals) {
+            if (is.null(param_vals$early_stopping_rounds)) {
+              stop("Parameter 'early_stopping_rounds' must be set to use internal tuning.")
+            }
+            assert_integerish(domain$upper, len = 1L, any.missing = FALSE)
+          }, .parent = topenv()),
+          disable_in_tune = list(early_stopping_rounds = NULL)
+        )
       )
       ps$add_dep(
         id = "mvs_reg", on = "bootstrap_type",
@@ -173,7 +194,13 @@ LearnerClassifCatboost = R6Class("LearnerClassifCatboost",
         predict_types = c("response", "prob"),
         param_set = ps,
         properties = c(
-          "missings", "weights", "importance", "twoclass", "multiclass"), # FIXME: parallel
+          "missings",
+          "weights",
+          "importance",
+          "twoclass",
+          "multiclass",
+          "internal_tuning",
+          "validation"),
         man = "mlr3extralearners::mlr_learners_classif.catboost",
         label = "Gradient Boosting"
       )
@@ -196,7 +223,10 @@ LearnerClassifCatboost = R6Class("LearnerClassifCatboost",
   ),
 
   private = list(
+    .validate = NULL,
+
     .train = function(task) {
+      pars = self$param_set$get_values(tags = "train")
 
       if (packageVersion("catboost") < "0.21") {
         stop("catboost v0.21 or greater is required.")
@@ -205,12 +235,11 @@ LearnerClassifCatboost = R6Class("LearnerClassifCatboost",
       # target is encoded as integer values from 0
       # if binary, the positive class is 1
       is_binary = length(task$class_names) == 2L
+      truth = task$data(cols = task$target_names)[[1L]]
       label = if (is_binary) {
-        ifelse(task$data(cols = task$target_names)[[1L]] == task$positive,
-          yes = 1L,
-          no = 0L)
+        ifelse(truth == task$positive, 1L, 0L)
       } else {
-        as.integer(task$data(cols = task$target_names)[[1L]]) - 1L
+        as.integer(truth) - 1L
       }
 
       # data must be a dataframe
@@ -220,8 +249,29 @@ LearnerClassifCatboost = R6Class("LearnerClassifCatboost",
         weight = task$weights$weight,
         thread_count = self$param_set$values$thread_count)
 
+      # early stopping
+      internal_valid_task = task$internal_valid_task
+
+      if (!is.null(pars$early_stopping_rounds) && is.null(internal_valid_task)) {
+        stopf("Learner (%s): Configure field 'validate' to enable early stopping.", self$id)
+      }
+
+      test_pool = if (!is.null(internal_valid_task)) {
+        # create test labels
+        test_label = if (length(internal_valid_task$class_names) == 2L) {
+          ifelse(internal_valid_task$data(cols = internal_valid_task$target_names)[[1L]] == internal_valid_task$positive, 1L, 0L)
+        } else {
+          as.integer(internal_valid_task$data(cols = internal_valid_task$target_names)[[1L]]) - 1L
+        }
+
+        invoke(catboost::catboost.load_pool,
+          data = internal_valid_task$data(cols = internal_valid_task$feature_names),
+          label = test_label,
+          weight = internal_valid_task$weights$weight,
+          thread_count = 1)
+      }
+
       # set loss_function correctly
-      pars = self$param_set$get_values(tags = "train")
       pars$loss_function = if (is_binary) {
         pars$loss_function_twoclass
       } else {
@@ -230,17 +280,19 @@ LearnerClassifCatboost = R6Class("LearnerClassifCatboost",
       pars$loss_function_twoclass = NULL
       pars$loss_function_multiclass = NULL
 
-      catboost::catboost.train(learn_pool, NULL, pars)
+      catboost::catboost.train(learn_pool, test_pool, pars)
     },
 
     .predict = function(task) {
-
       is_binary = (length(task$class_names) == 2L)
 
       # data must be a dataframe
       pool = invoke(catboost::catboost.load_pool,
         data = ordered_features(task, self),
         thread_count = self$param_set$values$thread_count)
+
+      # mlr3 ensures that the levels are correct
+      class_names = levels(task$truth(task$row_ids[1L])[[1L]])
 
       prediction_type = if (self$predict_type == "response") {
         "Class"
@@ -257,7 +309,7 @@ LearnerClassifCatboost = R6Class("LearnerClassifCatboost",
         response = if (is_binary) {
           ifelse(preds == 1L, yes = task$positive, no = task$negative)
         } else {
-          task$class_names[preds + 1L]
+          class_names[preds + 1L]
         }
         list(response = as.character(unname(response)))
       } else {
@@ -266,11 +318,46 @@ LearnerClassifCatboost = R6Class("LearnerClassifCatboost",
           preds = matrix(c(preds, 1 - preds), ncol = 2L, nrow = length(preds))
           colnames(preds) = c(task$positive, task$negative)
         } else {
-          colnames(preds) = self$state$train_task$class_names
+          colnames(preds) = class_names
         }
 
         list(prob = preds)
       }
+    },
+
+    .extract_internal_tuned_values = function() {
+      if (is.null(self$state$param_vals$early_stopping_rounds)) {
+        return(named_list())
+      }
+      list(iterations = self$model$tree_count)
+    },
+
+    .extract_internal_valid_scores = function() {
+      named_list()
+    }
+  ),
+
+  active = list(
+    #' @field internal_valid_scores
+    #' The last observation of the validation scores for all metrics.
+    #' Extracted from `model$evaluation_log`
+    internal_valid_scores = function() {
+      self$state$internal_valid_scores
+    },
+
+    #' @field internal_tuned_values
+    #' Returns the early stopped iterations if `early_stopping_rounds` was set during training.
+    internal_tuned_values = function() {
+      self$state$internal_tuned_values
+    },
+
+    #' @field validate
+    #' How to construct the internal validation data. This parameter can be either `NULL`, a ratio, `"test"`, or `"predefined"`.
+    validate = function(rhs) {
+      if (!missing(rhs)) {
+        private$.validate = mlr3::assert_validate(rhs)
+      }
+      private$.validate
     }
   )
 )